Mostly finished question edits.
This comprises of two large changes:
First, the _edit_question partial is cleaned up to cover 'fresh' questions and
editing existing questions. Second is the questions_controller which handles this.

If the question is new (Question.new), then Rails automatically chooses to do POST to /questions.
This corresponds to questions#create.

If the question exists (form_for q...) then Rails uses PATCH to /question/:id which goes
to the questions#update controller.

Note there is quite messy javascript to accomplish some of this. I feel like this is necessary
to be able to have the form feel more interactive (less page refreshes for simple commands like add questions).