Implemented updatable answers.
This was fairly easy thanks to the new answer_options model and strong parameters.

The slightly tricky bit involved answer_texts. This is either user-entered or automatically
generated based on the option_choice. Hence, a method exists in AnswerOption to populate this value
when needed.

Also, an option_id of 0 now denotes "unchecked", which will destroy an answer option when this value has been unset.