Refactored models to be "NotDeleteable".
What this means, is that a standard object.destroy will not actually remove the
db record, but set a flag deleted: true. These will then essentially be ignore by default.

On the other hand, a object.delete call will indeed remove the record, and any dependents.

This is good for undo-ing deletes etc.