Begun to modify website theme and styling.
Created a new css file theme.css.scss which contains all of the styling.
Currently, it basically works off the assumption that you define $theme-main
and tries to make everything work off of that. (This should be a dark-ish colour)

Made changes to the header and removed the footer for now.