Multiple changes: survey stats and the Tour page.
Survey stats have been updated a little to look a bit cleaner, and the
surveys index page is changed a little to show completed surveys.

Also, created a "tour" page, which highlights the main features of
the website (this is mostly for Sam to present to Norman, however, could
be useful in the future anyway).