Potentially finished refactoring.
Really cool side-effect. The logic for editing questions is SOOO much cleaner.
OptionGroups and OptionChoices are nested attributes of Questions and can be set in
forms as such. Therefore, the logic in questions controller can use strong
parameters properly. It even works nicely for deleting option choices with the
special field :_destroy.

Good stuff!

To summarise the changes: QuestionOption is no longer used.
Each question has ONE OptionGroup, and vice-versa. This is used as a catch-all for OptionChoices.
OptionGroups cannot be re-used, but there is a helper method OptionGroup.deep_copy
which will clone a question and the options for easy re-use.