Moved to acts_as_paranoid gem and fix #12 and fix #13
This replaces NotDeleteable, and has much the same impact, but
is slightly more sophisticated and introduces a
Model.with_deleted function which is nice to use.

This also closes #12 which was stupidly introduced.
Upgrading to Rails 4.0.3 to get around the bug where params can't be
empty arrays is a better fix than 7200f38dbefa3258be07dcc5568476e7a8406e02.